Implement a priority queue that allows for the user to get a suggestion as to who should flip the coin first (based on who has not recently flipped).
Priority queue works by adding the name of the flipper into the priority queue after their flip. The first name in the queue should be be the next one to go and the last name should be the last to go. Priority queue is saved in sharedPrefs and filters out non-existent children.