Redo logic for coin flip randomization, more refactoring
The way in which the coin flip worked was not logically sound in terms of hardware, no matter the workarounds as it will still impede the 50/50 nature of the flip. Now the coin will now flip for a set # of times instead of mathematically calculating how many flips it should do based on timing of the animation.